% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableOCR.R
\name{table_ocr}
\alias{table_ocr}
\title{TableOCR Shiny app}
\usage{
table_ocr(auth_file_path)
}
\arguments{
\item{auth_file_path}{Path to Google Vision API json credential file}
}
\description{
TableOCR Shiny app
}
