% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tableOCR-package}
\alias{tableOCR}
\alias{tableOCR-package}
\title{tableOCR: OCR Table data}
\description{
This package contains functions we frequently use.
}
\author{
\strong{Maintainer}: Thomas Tichelbaecker \email{tt9@princeton.edu}

}
\keyword{internal}
