% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableOCR.R
\name{table_ocr}
\alias{table_ocr}
\title{Fahrenheit conversion}
\usage{
table_ocr(auth_file_path)
}
\arguments{
\item{F_temp}{The temperature in degrees Fahrenheit}
}
\value{
The temperature in degrees Celsius
}
\description{
Convert degrees Fahrenheit temperatures to degrees Celsius
}
\examples{
temp1 <- F_to_C(50);
temp2 <- F_to_C( c(50, 63, 23) );
}
